+++ src/containers/Routes.jsx
@@ -5,7 +5,7 @@ import { HashRouter, Route, Switch } from 'react-router-dom';
 import Home from './Home';
 import Quiz from './Quiz';
 import Login from './Login';
-import Signup from './Signup';
+import Signup from './signup/Signup';
 import Nav from '../components/nav/Nav';
 
 const Routes = () => (
