+++ src/containers/signup/Signup.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 
-import SignupForm from '../components/forms/SignupForm';
-import { userSignupRequest } from '../actions/user';
+import SignupForm from './SignupForm';
+import { userSignupRequest } from '../../actions/user';
 
 class Signup extends React.Component {
   submit = (vals) => {
