+++ src/containers/signup/SignupForm.jsx
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { reduxForm } from 'redux-form';
 
-import UserForm from './UserForm';
+import UserForm from '../../components/forms/UserForm';
 
 const mapStateToProps = state => ({
   showErr: state.errors.onSignup,
